Further cleanups and refactorings
