+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/algorithms/images/ImageUtility.cs
@@ -2,13 +2,13 @@
 
 namespace com.csutil.algorithms.images {
 
-    public class ImageUtility {
+    public static class ImageUtility {
 
-        public static byte[] FlipImageHorizontally(ImageResult sourceImage) {
+        public static byte[] FlipImageHorizontally(this ImageResult sourceImage) {
             return FlipImageHorizontally(sourceImage.Data, sourceImage.Width, sourceImage.Height, (int)sourceImage.ColorComponents);
         }
-        
-        public static byte[] FlipImageHorizontally(byte[] imageData, int width, int height, int bytesPerPixel) {
+
+        private static byte[] FlipImageHorizontally(byte[] imageData, int width, int height, int bytesPerPixel) {
             var flippedImage = new byte[imageData.Length];
             for (var y = 0; y < height; y++) {
                 for (var x = 0; x < width; x++) {
@@ -22,7 +22,7 @@ namespace com.csutil.algorithms.images {
             return flippedImage;
         }
 
-        public static byte[] FlipImageVertically(ImageResult sourceImage) {
+        public static byte[] FlipImageVertically(this ImageResult sourceImage) {
             return FlipImageVertically(sourceImage.Data, sourceImage.Width, sourceImage.Height, (int)sourceImage.ColorComponents);
         }
 
@@ -40,19 +40,11 @@ namespace com.csutil.algorithms.images {
             return flippedImage;
         }
 
-        public static byte[] FlipImageVerticallyVariant2(byte[] imageData, int width, int height, int bytesPerPixel = 4) {
-            byte[] flippedImage = new byte[imageData.Length];
-            for (int y = 0; y < height; y++) {
-                for (int x = 0; x < width * bytesPerPixel; x++) {
-                    var originalIndex = y * width * bytesPerPixel + x;
-                    var flippedIndex = (height - 1 - y) * width * bytesPerPixel + x;
-                    flippedImage[flippedIndex] = imageData[originalIndex];
-                }
-            }
-            return flippedImage;
+        public static byte[] CroppingImage(this ImageResult sourceImage, int cropX, int cropY, int cropWidth, int cropHeight) {
+            return CroppingImage(sourceImage.Data, sourceImage.Width, sourceImage.Height, (int)sourceImage.ColorComponents, cropX, cropY, cropWidth, cropHeight);
         }
-
-        public static byte[] CroppingImage(byte[] image, int originalWidth, int originalHeight, int bytesPerPixel, int cropX, int cropY, int cropWidth, int cropHeight) {
+        
+        private static byte[] CroppingImage(byte[] image, int originalWidth, int originalHeight, int bytesPerPixel, int cropX, int cropY, int cropWidth, int cropHeight) {
             var croppedImage = new byte[cropWidth * cropHeight * bytesPerPixel];
 
             for (var y = 0; y < cropHeight; y++) {
@@ -65,11 +57,14 @@ namespace com.csutil.algorithms.images {
                     }
                 }
             }
-
             return croppedImage;
         }
 
-        public static byte[] ResizeImage(byte[] originalImage, int originalWidth, int originalHeight, int bytesPerPixel, int newWidth, int newHeight) {
+        public static byte[] ResizeImage(this ImageResult sourceImage, int newWidth, int newHeight) {
+            return ResizeImage(sourceImage.Data, sourceImage.Width, sourceImage.Height, (int)sourceImage.ColorComponents, newWidth, newHeight);
+        }
+        
+        private static byte[] ResizeImage(byte[] originalImage, int originalWidth, int originalHeight, int bytesPerPixel, int newWidth, int newHeight) {
             byte[] resizedImage = new byte[newWidth * newHeight * bytesPerPixel];
 
             double xRatio = (double)originalWidth / newWidth;
@@ -88,9 +83,9 @@ namespace com.csutil.algorithms.images {
                     }
                 }
             }
-
             return resizedImage;
         }
 
     }
+
 }
