+++ CsCore/PlainNetClassLib/src/Plugins/CsCore/com/csutil/io/ImageFileWriter.cs
@@ -8,12 +8,12 @@ namespace com.csutil.io {
     public static class ImageFileWriter {
 
         public static void WritePngToStream(this ImageResult sourceImage, Stream targetStream) {
-            var verticallyFlipped = ImageUtility.FlipImageVertically(sourceImage);
+            var verticallyFlipped = sourceImage.FlipImageVertically();
             new ImageWriter().WritePng(verticallyFlipped, sourceImage.Width, sourceImage.Height, StbImageWriteSharp.ColorComponents.RedGreenBlueAlpha, targetStream);
         }
 
         public static void WriteJpgToStream(this ImageResult sourceImage, Stream targetStream, int quality) {
-            var verticallyFlipped = ImageUtility.FlipImageVertically(sourceImage);
+            var verticallyFlipped = sourceImage.FlipImageVertically();
             new ImageWriter().WriteJpg(verticallyFlipped, sourceImage.Width, sourceImage.Height, StbImageWriteSharp.ColorComponents.RedGreenBlueAlpha, targetStream, quality);
         }
